// src/app/page/task/redux/model/task-action-type.ts
enum TaskActionType {
    CHANGE_POSITION = 'CHANGE_POSITION',
    CHANGE_SOLUTION = 'CHANGE_SOLUTION',
    SUBMIT_TASK_START = 'SUBMIT_TASK_START',
    SUBMIT_TASK_SUCCESS = 'SUBMIT_TASK_SUCCESS',
    SUBMIT_TASK_FAILURE = 'SUBMIT_TASK_FAILURE',
    GO_TO_NEXT_STEP = 'GO_TO_NEXT_STEP',
    GO_TO_PREVIOUS_STEP = 'GO_TO_PREVIOUS_STEP',
}

export {TaskActionType}